requirements.txt
black==22.3.0
django==3.2.13
djangorestframework==3.13.1
drf-spectacular==0.22.1
flake8==4.0.1
pre-commit==2.19.0
psycopg2==2.9.3
pylint==2.14.3
urllib3==1.26.4
